% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{gen_semicontinuous_data}
\alias{gen_semicontinuous_data}
\title{Generates data from a two part distribution with a point mass at zero and heterogeneous treatment effects}
\usage{
gen_semicontinuous_data(n.obs = 1000, n.vars = 25)
}
\arguments{
\item{n.obs}{number of observations}

\item{n.vars}{number of variables. Must be at least 10}
}
\value{
returns list with values \code{y} for outcome, \code{x} for design matrix, \code{trt} for
treatment assignments, \code{betanonzero} for true coefficients for treatment-covariate interactions for model for
whether or not a response is nonzero, \code{betapos} for true coefficients for treatment-covariate interactions
for positive model, \code{treatment_risk_ratio} for the true covariate-conditional treatment effect risk ratio for
each observation, \code{pi.x} for the true underlying propensity score
}
\description{
Generates semicontinuous data with HTE
}
